classes for animation and messages, some fixes
